# packages/shared/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExpenseEntryCreate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "budget_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Id"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "entry_date": {
            "format": "date",
            "title": "Entry Date",
            "type": "string"
          },
          "is_planned": {
            "default": false,
            "title": "Is Planned",
            "type": "boolean"
          },
          "is_recurring": {
            "default": false,
            "title": "Is Recurring",
            "type": "boolean"
          },
          "label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "merchant": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recurrence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecurrenceRuleCreate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "label",
          "amount",
          "entry_date"
        ],
        "title": "ExpenseEntryCreate",
        "type": "object"
      },
      "ExpenseEntryRead": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "budget_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Id"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "entry_date": {
            "format": "date",
            "title": "Entry Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_planned": {
            "title": "Is Planned",
            "type": "boolean"
          },
          "is_recurring": {
            "title": "Is Recurring",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recurrence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecurrenceRuleRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "recurrence_rule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Rule Id"
          },
          "source_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Amount"
          },
          "source_entry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Entry Date"
          }
        },
        "required": [
          "id",
          "label",
          "amount",
          "entry_date",
          "category_id",
          "is_recurring",
          "recurrence_rule_id",
          "merchant",
          "notes",
          "is_planned",
          "budget_id"
        ],
        "title": "ExpenseEntryRead",
        "type": "object"
      },
      "ExpenseEntryUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "budget_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Id"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "entry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Date"
          },
          "is_planned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Planned"
          },
          "is_recurring": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Recurring"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "merchant": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recurrence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecurrenceRuleCreate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ExpenseEntryUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecurrenceFrequency": {
        "enum": [
          "weekly",
          "monthly",
          "yearly"
        ],
        "title": "RecurrenceFrequency",
        "type": "string"
      },
      "RecurrenceRuleCreate": {
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency": {
            "$ref": "#/components/schemas/RecurrenceFrequency"
          },
          "interval_count": {
            "default": 1,
            "maximum": 52.0,
            "minimum": 1.0,
            "title": "Interval Count",
            "type": "integer"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "frequency",
          "start_date"
        ],
        "title": "RecurrenceRuleCreate",
        "type": "object"
      },
      "RecurrenceRuleRead": {
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency": {
            "$ref": "#/components/schemas/RecurrenceFrequency"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interval_count": {
            "title": "Interval Count",
            "type": "integer"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "frequency",
          "interval_count",
          "start_date",
          "end_date"
        ],
        "title": "RecurrenceRuleRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "gcol_session_v2",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Germany Cost of Living API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/expenses": {
      "get": {
        "operationId": "list_expenses_api_expenses_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExpenseEntryRead"
                  },
                  "title": "Response List Expenses Api Expenses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Expenses",
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "operationId": "create_expense_api_expenses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseEntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Expense",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/expenses/{entry_id}": {
      "delete": {
        "operationId": "delete_expense_api_expenses__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Expense",
        "tags": [
          "expenses"
        ]
      },
      "patch": {
        "operationId": "update_expense_api_expenses__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseEntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseEntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Update Expense",
        "tags": [
          "expenses"
        ]
      }
    },
...
  }
}